requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt[crypto]
drf-yasg
django-cors-headers
django-filter
Pillow
